+++ engine/src/main/kotlin/world/gregs/voidps/engine/timer/Ticks.kt
@@ -3,13 +3,13 @@ package world.gregs.voidps.engine.timer
 import java.util.concurrent.TimeUnit
 
 class TickTime {
-    fun toNanos(duration: Long) = TimeUnit.MILLISECONDS.toNanos(toMillis(duration))
-    fun toMicros(duration: Long) = TimeUnit.MILLISECONDS.toMicros(toMillis(duration))
-    fun toMillis(duration: Long) = duration * 600
-    fun toSeconds(duration: Long) = TimeUnit.MILLISECONDS.toSeconds(toMillis(duration))
-    fun toMinutes(duration: Long) = TimeUnit.MILLISECONDS.toMinutes(toMillis(duration))
-    fun toHours(duration: Long) = TimeUnit.MILLISECONDS.toHours(toMillis(duration))
-    fun toDays(duration: Long) = TimeUnit.MILLISECONDS.toDays(toMillis(duration))
+    fun toNanos(duration: Int) = TimeUnit.MILLISECONDS.toNanos(toMillis(duration))
+    fun toMicros(duration: Int) = TimeUnit.MILLISECONDS.toMicros(toMillis(duration))
+    fun toMillis(duration: Int) = duration * 600L
+    fun toSeconds(duration: Int) = TimeUnit.MILLISECONDS.toSeconds(toMillis(duration))
+    fun toMinutes(duration: Int) = TimeUnit.MILLISECONDS.toMinutes(toMillis(duration))
+    fun toHours(duration: Int) = TimeUnit.MILLISECONDS.toHours(toMillis(duration))
+    fun toDays(duration: Int) = TimeUnit.MILLISECONDS.toDays(toMillis(duration))
     fun toClientTicks(duration: Int) = duration * 30
 }
